% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mesh.3D}
\alias{plot.mesh.3D}
\title{Plot a mesh.3D object}
\usage{
\method{plot}{mesh.3D}(x, ...)
}
\arguments{
\item{x}{A \code{mesh.3D} object generated by \code{create.mesh.3D} or \code{second.order.mesh.3D}.}

\item{...}{Arguments representing graphical options to be passed to \link[graphics]{par}.}
}
\description{
Plot the triangulation of a mesh \code{mesh.3D} object, generated by \code{create.mesh.3D}. The points indicates the mesh nodes.
}
\examples{
##Load a mesh.3D object
data(sphere3D)
##Plot the triangulation of the object
plot.mesh.3D(sphere3D)
}
